requirements.txt
fastapi
pydantic
openai
python-dateutil
glob2
uvicorn
requests
faker
pillow
